Need to add doc and fix benchmark for skip stuff, as this introduces a regression.

Before

`test postings::tests::bench_skip_next_p01                  ... bench:      98,518 ns/iter (+/- 2,588)
test postings::tests::bench_skip_next_p1                   ... bench:     349,363 ns/iter (+/- 22,189)
test postings::tests::bench_skip_next_p10                  ... bench:     694,869 ns/iter (+/- 18,959)
test postings::tests::bench_skip_next_p90                  ... bench:     475,763 ns/iter (+/- 41,806)`

After 

`test postings::tests::bench_skip_next_p01                  ... bench:     118,199 ns/iter (+/- 21,571)
test postings::tests::bench_skip_next_p1                   ... bench:     583,155 ns/iter (+/- 71,572)
test postings::tests::bench_skip_next_p10                  ... bench:   1,745,838 ns/iter (+/- 314,828)
test postings::tests::bench_skip_next_p90                  ... bench:     825,213 ns/iter (+/- 139,262)`