@fulmicoton 2017-05-20
perf following @lnicola 's fix

`test postings::tests::bench_skip_next_p01                  ... bench:     118,797 ns/iter (+/- 68,724)
test postings::tests::bench_skip_next_p1                   ... bench:     522,549 ns/iter (+/- 911,815)
test postings::tests::bench_skip_next_p10                  ... bench:   1,092,403 ns/iter (+/- 296,544)
test postings::tests::bench_skip_next_p90                  ... bench:     788,155 ns/iter (+/- 140,590)`