@lnicola 2017-05-20
Now this is awkward. After my last commit, on my Linux box I get:

```
master:
test postings::tests::bench_segment_intersection           ... bench:   5,850,617 ns/iter (+/- 191,407)
test postings::tests::bench_segment_postings               ... bench:   1,969,299 ns/iter (+/- 8,332)
test postings::tests::bench_skip_next_p01                  ... bench:     279,834 ns/iter (+/- 4,155)
test postings::tests::bench_skip_next_p1                   ... bench:   1,024,438 ns/iter (+/- 25,858)
test postings::tests::bench_skip_next_p10                  ... bench:   2,707,760 ns/iter (+/- 36,008)
test postings::tests::bench_skip_next_p90                  ... bench:   2,869,197 ns/iter (+/- 48,216)

pr:
test postings::tests::bench_segment_intersection           ... bench:   4,288,185 ns/iter (+/- 77,691)
test postings::tests::bench_segment_postings               ... bench:     902,715 ns/iter (+/- 6,661)
test postings::tests::bench_skip_next_p01                  ... bench:     256,476 ns/iter (+/- 1,006)
test postings::tests::bench_skip_next_p1                   ... bench:     882,680 ns/iter (+/- 12,647)
test postings::tests::bench_skip_next_p10                  ... bench:   2,134,069 ns/iter (+/- 22,092)
test postings::tests::bench_skip_next_p90                  ... bench:   2,071,681 ns/iter (+/- 37,390)
```

But on my Windows laptop:

```
master:

pr:
```