@lnicola 2017-05-20
Also, why are my Windows times worse than those reported in https://github.com/tantivy-search/tantivy/pull/153?