@fulmicoton 2017-05-21
After the last commit

`test postings::tests::bench_skip_next_p01                  ... bench:     114,654 ns/iter (+/- 27,259)
test postings::tests::bench_skip_next_p1                   ... bench:     471,854 ns/iter (+/- 85,904)
test postings::tests::bench_skip_next_p10                  ... bench:   1,114,490 ns/iter (+/- 494,509)
test postings::tests::bench_skip_next_p90                  ... bench:     702,411 ns/iter (+/- 184,332)`