@fulmicoton 2017-05-21
```
test postings::tests::bench_skip_next_p01                  ... bench:     108,684 ns/iter (+/- 16,101)
test postings::tests::bench_skip_next_p1                   ... bench:     443,134 ns/iter (+/- 59,802)
test postings::tests::bench_skip_next_p10                  ... bench:     973,382 ns/iter (+/- 145,465)
test postings::tests::bench_skip_next_p90                  ... bench:     575,853 ns/iter (+/- 100,085)
```