@lnicola 2017-05-21
```
test postings::tests::bench_segment_intersection           ... bench:   1,551,921 ns/iter (+/- 146,069)
test postings::tests::bench_segment_postings               ... bench:     220,643 ns/iter (+/- 24,668)
test postings::tests::bench_skip_next_p01                  ... bench:     106,301 ns/iter (+/- 4,230)
test postings::tests::bench_skip_next_p1                   ... bench:     402,466 ns/iter (+/- 34,051)
test postings::tests::bench_skip_next_p10                  ... bench:     840,000 ns/iter (+/- 77,473)
test postings::tests::bench_skip_next_p90                  ... bench:     467,733 ns/iter (+/- 61,548)
```