@lnicola 2017-05-21
```
test postings::tests::bench_segment_intersection           ... bench:   1,447,625 ns/iter (+/- 122,742)
test postings::tests::bench_segment_postings               ... bench:     206,747 ns/iter (+/- 19,893)
test postings::tests::bench_skip_next_p01                  ... bench:      99,848 ns/iter (+/- 7,773)
test postings::tests::bench_skip_next_p1                   ... bench:     374,617 ns/iter (+/- 36,732)
test postings::tests::bench_skip_next_p10                  ... bench:     840,755 ns/iter (+/- 53,914)
test postings::tests::bench_skip_next_p90                  ... bench:     475,140 ns/iter (+/- 60,780)
```